requirements.txt
pygame-ce
cryptography
numpy
filterpy
pymavlink
pyserial
geopy
python-dotenv
label-studio
opencv-python
pyproj
gymnasium
aiortc
aiohttp
ruff
